Yet Another™ semantic overhaul to how the DisplayDevice works. All state now resides in m_context instead, simplifying move semantics considerably. Default-enabled Settings menu tab, with toggle for the Debug tab. Debugging like so forces the display to always reference the metadata_staging struct, rather than what each core passes along in a frame, allowing for real-time screwing around. Texture zoom renamed to Minimum zoom. Still need to mess with the Metadata struct and make it proper atomic for further safety, and also swap to the upcoming Parameter-style type semantics for better ergonomics.
FrontendInterface menu tab handling now performs pre-pass to clear stale registrations prior to producing a menu tab, preventing stale tabs that go away as soon as they're clicked. The autohide_menubar now instead behaves more like a toggle_menubar instead for simplicity, but still hides if window loses focus.

+++ include/systems/CHIP8/Chip8_CoreInterface.cpp
@@ -27,7 +27,7 @@ Chip8_CoreInterface::Chip8_CoreInterface(DisplayDevice display_device) noexcept
 
 	mDisplayDevice.set_osd_callable([&]() {
 		if (mInterrupt == Interrupt::INPUT) {
-			osd::key_press_indicator(WaveForms::pulse_t(500, Millis::now()).as_unipolar());
+			osd::key_press_indicator(WaveForms::pulse_t(500, u32(Millis::now())).as_unipolar());
 		}
 		if (has_system_state(EmuState::STATS)) {
 			osd::simple_text_overlay(copy_statistics_string());
